Fixing pre-register page
