+++ src/pages/PreRegister/styles.js
@@ -15,18 +15,6 @@ export const Body = styled.div`
     flex-direction: column;
 `;
 
-export const BackIcon = styled.img`
-    width: 3vw;
-    height: 4vh;
-    margin: 10px 0px 0px 10px;
-`;
-
-export const BackLink = styled(Link)`
-    width: 3vw;
-    height: 4vh;
-    margin: 10px 0px 0px 10px;
-`;
-
 export const Title = styled.p`
     margin: 3vh auto;
     display: inline;
@@ -90,7 +78,7 @@ export const FieldDiv = styled.div`
     height: 12vh;
     margin-left: 20px;
 `;
-    
+
 export const FieldName = styled.label`
     float: left;
     margin: 10px 0px 5px 10px;
@@ -101,7 +89,7 @@ export const FieldName = styled.label`
     line-height: 24px;
     color: #000000;
 `;
-    
+
 export const Input = styled.input`
     display: block;
     height: 5vh;
@@ -194,7 +182,7 @@ export const SendButton = styled.button`
     text-decoration: none;
     cursor: pointer;
 `;
-    
+
 export const SendButtonName = styled.p`
     text-align: center;
     text-decoration: none;
