+++ src/sharedComponents/BackIcon/styles.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+import { Link } from 'react-router-dom'
+
+export const BackIcon = styled.img`
+    width: 3vw;
+    height: 4vh;
+    margin: 10px 0px 0px 10px;
+`;
+
+export const BackLink = styled(Link)`
+    width: 3vw;
+    height: 4vh;
+    margin: 10px 0px 0px 10px;
+`;
